Coordinates + camera working

+++ src/components/camera.rs
@@ -7,7 +7,7 @@ pub const MAX_CAMERA_RENDER_AREA_Y: usize = 3;
 
 #[derive(Component)]
 pub struct CameraComponent {
-    pub speed: f32,
+    pub speed: f64,
     pub visible_layer: u32,
-    pub zoom: f32,
+    pub zoom: f64,
 }
